refactor: extract worker ops into diesel_common
