+++ crates/taskito-core/src/storage/diesel_common/mod.rs
@@ -10,6 +10,7 @@ mod locks;
 mod logs;
 mod metrics;
 pub(crate) mod migrations;
+mod workers;
 
 pub(crate) use archival::impl_diesel_archival_ops;
 pub(crate) use dead_letter::impl_diesel_dead_letter_ops;
@@ -17,3 +18,4 @@ pub(crate) use jobs::impl_diesel_job_ops;
 pub(crate) use locks::impl_diesel_lock_ops;
 pub(crate) use logs::impl_diesel_log_ops;
 pub(crate) use metrics::impl_diesel_metric_ops;
+pub(crate) use workers::impl_diesel_worker_ops;
